# docs/swagger.json
{
    "schemes": [
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Happy integration",
        "title": "CodeNexus API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
        "/api/v1/auth/signin": {
            "post": {
                "description": "Sign in a user with the provided credentials",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Sign in a user",
                "parameters": [
                    {
                        "description": "User login details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Signin successfully",
                        "schema": {
                            "$ref": "#/definitions/dto.TLoginResponse"
                        }
                    },
                    "400": {
                        "description": "invalid payload\" or \"error validating payload",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "user not found\" or \"your account has not been verified\" or \"incorrect password",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/signup": {
            "post": {
                "description": "Register a new user with the provided details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Register a new user",
                "parameters": [
                    {
                        "description": "User registration details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "registration is successful, please check your email for email verification",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid payload\" or \"error validating payload\" or \"registration failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/auth/verify": {
            "post": {
                "description": "Verify user email with the provided token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Verify user email",
                "parameters": [
                    {
                        "description": "Email verification details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TVerificationRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Your email has been successfully verified",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "invalid payload\" or \"error validating payload\" or \"email verification failed",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/me": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Get information about the currently authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get current user",
                "responses": {
                    "200": {
                        "description": "User information",
                        "schema": {
                            "$ref": "#/definitions/dto.TGetUserResponse"
                        }
                    },
                    "401": {
                        "description": "user not found\" or \"Access denied",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/me/avatar": {
            "patch": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Update the avatar of the currently authenticated user",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Update user avatar",
                "parameters": [
                    {
                        "type": "file",
                        "description": "Image file to upload",
                        "name": "avatar",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Avatar update success",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request or error uploading image",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "User not found or unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/v1/user/me/detail": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Create user detail for the currently authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Create user detail",
                "parameters": [
                    {
                        "description": "User detail information",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.TCreateUserDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User detail created successfully",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid input format",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "User not found\" or \"Access denied",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.TCreateUserDetailRequest": {
            "type": "object",
            "properties": {
                "address": {
                    "description": "Address of the user (optional)\nExample: 123 Main Street",
                    "type": "string"
                },
                "job": {
                    "description": "Job of the user (optional)\nExample: Software Engineer",
                    "type": "string"
                },
                "phone": {
                    "description": "Phone number of the user (optional)\nExample: +1234567890",
                    "type": "string"
                }
            }
        },
        "dto.TGetUserResponse": {
            "type": "object",
            "properties": {
                "avatar": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "is_verified": {
                    "type": "boolean"
                },
                "role": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "dto.TLoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "Email of the user (required, should be a valid email address)\nExample: john.doe@example.com\nRequired: true\nFormat: email",
                    "type": "string"
                },
                "password": {
                    "description": "Password of the user (required, minimum length 8)\nExample: mySecurePassword\nRequired: true\nMinLength: 8",
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "dto.TLoginResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                }
            }
        },
        "dto.TRegisterRequest": {
            "type": "object",
            "required": [
                "email",
                "name",
                "password"
            ],
            "properties": {
                "email": {
                    "description": "Email of the user (required, should be a valid email address)\nExample: john.doe@example.com\nRequired: true\nFormat: email",
                    "type": "string"
                },
                "name": {
                    "description": "Name of the user (required)\nExample: John Doe\nRequired: true",
                    "type": "string"
                },
                "password": {
                    "description": "Password of the user (required, minimum length 8)\nExample: mySecurePassword\nRequired: true\nMinLength: 8",
                    "type": "string",
                    "minLength": 8
                }
            }
        },
        "dto.TVerificationRequest": {
            "type": "object",
            "required": [
                "email",
                "token"
            ],
            "properties": {
                "email": {
                    "description": "Email of the user (required, should be a valid email address)\nExample: john.doe@example.com\nRequired: true\nFormat: email",
                    "type": "string"
                },
                "token": {
                    "description": "Token for the user (required)\nExample: 1HV3O4\nRequired: true\nMinLength: 6",
                    "type": "string",
                    "minLength": 6
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}